# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Evidence-led IVF clinic directory API for fictional demonstration data.
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Location {
  slug: string;
  city: string;
  state: string;
  /** @minimum 0 */
  clinicCount: number;
}

export interface Service {
  id: string;
  name: string;
  slug: string;
  /** @nullable */
  description?: string | null;
}

export interface ServiceInput {
  /** @minLength 1 */
  name: string;
  /** @minLength 1 */
  slug: string;
  /** @nullable */
  description?: string | null;
}

export interface Source {
  id: string;
  title: string;
  sourceType: string;
  /** @minLength 1 */
  url: string;
  /** @nullable */
  publisher?: string | null;
  /** @nullable */
  publishedOn?: string | null;
  /** @nullable */
  notes?: string | null;
}

export interface SourceInput {
  /** @minLength 1 */
  title: string;
  /** @minLength 1 */
  sourceType: string;
  /** @minLength 1 */
  url: string;
  /** @nullable */
  publisher?: string | null;
  /** @nullable */
  publishedOn?: string | null;
  /** @nullable */
  notes?: string | null;
}

export interface RateObservation {
  id: string;
  clinicId: string;
  outcomeType: string;
  outcomeDefinition: string;
  /**
     * @minimum 0
     * @maximum 100
     */
  ratePercentage: number;
  /**
     * @minimum 0
     * @nullable
     */
  numerator?: number | null;
  /**
     * @minimum 0
     * @nullable
     */
  denominatorCount?: number | null;
  denominatorType: string;
  denominatorDefinition: string;
  ageBand: string;
  /** @nullable */
  ageMeasurementPoint?: string | null;
  eggSource: string;
  treatmentContext: string;
  treatmentType: string;
  /** @nullable */
  priorTreatmentCohort?: string | null;
  cumulativeMethod: string;
  reportingPeriodStart: string;
  reportingPeriodEnd: string;
  yearLabel: string;
  methodologyNotes: string;
  smallSample: boolean;
  source: Source;
  verificationStatus: string;
  publicationStatus: string;
  comparabilityGroupKey: string;
}

export interface RateObservationInput {
  clinicId: string;
  /** @minLength 1 */
  outcomeType: string;
  /** @minLength 1 */
  outcomeDefinition: string;
  /**
     * @minimum 0
     * @maximum 100
     */
  ratePercentage: number;
  /**
     * @minimum 0
     * @nullable
     */
  numerator?: number | null;
  /**
     * @minimum 0
     * @nullable
     */
  denominatorCount?: number | null;
  /** @minLength 1 */
  denominatorType: string;
  /** @minLength 1 */
  denominatorDefinition: string;
  /** @minLength 1 */
  ageBand: string;
  /** @nullable */
  ageMeasurementPoint?: string | null;
  /** @minLength 1 */
  eggSource: string;
  /** @minLength 1 */
  treatmentContext: string;
  /** @minLength 1 */
  treatmentType: string;
  /** @nullable */
  priorTreatmentCohort?: string | null;
  /** @minLength 1 */
  cumulativeMethod: string;
  reportingPeriodStart: string;
  reportingPeriodEnd: string;
  /** @minLength 1 */
  yearLabel: string;
  /** @minLength 1 */
  methodologyNotes: string;
  smallSample?: boolean;
  /** @minLength 1 */
  sourceId: string;
  /** @minLength 1 */
  verificationStatus: string;
}

export interface Clinic {
  id: string;
  slug: string;
  name: string;
  city: string;
  state: string;
  /** @nullable */
  address?: string | null;
  /** @nullable */
  latitude?: number | null;
  /** @nullable */
  longitude?: number | null;
  /** @nullable */
  licensingStatus?: string | null;
  /** @nullable */
  regulator?: string | null;
  /** @nullable */
  phone?: string | null;
  /**
     * @minLength 1
     * @nullable
     */
  email?: string | null;
  /**
     * @minLength 1
     * @nullable
     */
  website?: string | null;
  services: Service[];
  headlineObservation?: RateObservation | null;
  lastReviewedAt: string;
  demonstrationData?: boolean;
}

export type ClinicProfile = Clinic & {
  observations: RateObservation[];
};

export interface ClinicListResponse {
  items: Clinic[];
  /** @minimum 0 */
  total: number;
  availableLocations: Location[];
  availableServices: Service[];
}

export interface ClinicInput {
  /** @minLength 1 */
  name: string;
  /** @minLength 1 */
  slug: string;
  /** @minLength 1 */
  city: string;
  /** @minLength 1 */
  state: string;
  /** @nullable */
  address?: string | null;
  /** @nullable */
  latitude?: number | null;
  /** @nullable */
  longitude?: number | null;
  /** @nullable */
  licensingStatus?: string | null;
  /** @nullable */
  regulator?: string | null;
  /** @nullable */
  phone?: string | null;
  /**
     * @minLength 1
     * @nullable
     */
  email?: string | null;
  /**
     * @minLength 1
     * @nullable
     */
  website?: string | null;
  serviceIds?: string[];
}

export type ClinicUpdate = ClinicInput;

export type AdminClinic = Clinic & {
  recordStatus: string;
};

export interface CorrectionInput {
  /** @minLength 1 */
  clinicSlug: string;
  /** @nullable */
  observationId?: string | null;
  /**
     * @minLength 10
     * @maxLength 5000
     */
  message: string;
  /** @minLength 1 */
  contactEmail: string;
}

export interface CorrectionReceipt {
  id: string;
  receivedAt: string;
}

/**
 * @nullable
 */
export type AuditEventBeforeSnapshot = { [key: string]: unknown } | null;

/**
 * @nullable
 */
export type AuditEventAfterSnapshot = { [key: string]: unknown } | null;

export interface AuditEvent {
  id: string;
  action: string;
  entityType: string;
  entityId: string;
  /** @nullable */
  beforeSnapshot: AuditEventBeforeSnapshot;
  /** @nullable */
  afterSnapshot: AuditEventAfterSnapshot;
  createdAt: string;
}

export interface AdminSummary {
  /** @minimum 0 */
  clinicCount: number;
  /** @minimum 0 */
  publishedObservationCount: number;
  /** @minimum 0 */
  draftObservationCount: number;
  /** @minimum 0 */
  pendingCorrectionCount: number;
  recentAuditEvents: AuditEvent[];
}

export type ImportPreviewInputMapping = {[key: string]: string};

export interface ImportPreviewInput {
  /** @minLength 1 */
  csvText: string;
  mapping: ImportPreviewInputMapping;
}

export interface ImportRowError {
  /** @minimum 1 */
  rowNumber: number;
  field: string;
  message: string;
}

export interface ImportPreview {
  /** @minimum 0 */
  totalRows: number;
  /** @minimum 0 */
  validRows: number;
  /** @minimum 0 */
  invalidRows: number;
  /** @minimum 0 */
  duplicateRows: number;
  errors: ImportRowError[];
}

export interface SecuritySession {
  userId: string;
  role: string;
}

export type SearchParameter = string;

export type LocationParameter = string;

export type ServiceParameter = string;

export type AgeBandParameter = string;

export type OutcomeTypeParameter = string;

export type DenominatorTypeParameter = string;

export type EggSourceParameter = string;

export type VerificationTypeParameter = string;

export type ReportingYearParameter = number;

export type ListClinicsParams = {
q?: SearchParameter;
location?: LocationParameter;
service?: ServiceParameter;
ageBand?: AgeBandParameter;
outcomeType?: OutcomeTypeParameter;
denominatorType?: DenominatorTypeParameter;
eggSource?: EggSourceParameter;
verificationType?: VerificationTypeParameter;
reportingYear?: ReportingYearParameter;
};