# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Evidence-led IVF clinic directory API for fictional demonstration data.
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /clinics:
    get:
      operationId: listClinics
      tags: [clinics]
      summary: List eligible published clinics
      parameters:
        - $ref: "#/components/parameters/Search"
        - $ref: "#/components/parameters/Location"
        - $ref: "#/components/parameters/Service"
        - $ref: "#/components/parameters/AgeBand"
        - $ref: "#/components/parameters/OutcomeType"
        - $ref: "#/components/parameters/DenominatorType"
        - $ref: "#/components/parameters/EggSource"
        - $ref: "#/components/parameters/VerificationType"
        - $ref: "#/components/parameters/ReportingYear"
      responses:
        "200":
          description: Published clinics matching the filters
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClinicListResponse"
  /clinics/{slug}:
    get:
      operationId: getClinic
      tags: [clinics]
      summary: Get a published clinic profile and eligible observations
      parameters:
        - $ref: "#/components/parameters/ClinicSlug"
      responses:
        "200":
          description: Clinic profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClinicProfile"
        "404":
          description: Clinic not found
...
components:
  schemas:
...
    Location:
      type: object
      required: [slug, city, state, clinicCount]
      properties:
        slug: { type: string }
        city: { type: string }
        state: { type: string }
        clinicCount: { type: number, minimum: 0, multipleOf: 1 }
    Service:
      type: object
      required: [id, name, slug]
      properties:
        id: { type: string }
        name: { type: string }
        slug: { type: string }
        description: { type: ["string", "null"] }
...
    Source:
      type: object
      required: [id, title, sourceType, url]
      properties:
        id: { type: string }
        title: { type: string }
        sourceType: { type: string }
        url: { type: string, minLength: 1 }
        publisher: { type: ["string", "null"] }
        publishedOn: { type: ["string", "null"], format: date }
        notes: { type: ["string", "null"] }
...
    RateObservation:
      type: object
      required:
        - id
        - clinicId
        - outcomeType
        - outcomeDefinition
        - ratePercentage
        - denominatorType
        - denominatorDefinition
        - ageBand
        - eggSource
        - treatmentContext
        - treatmentType
        - cumulativeMethod
        - reportingPeriodStart
        - reportingPeriodEnd
        - yearLabel
        - methodologyNotes
        - smallSample
        - source
        - verificationStatus
        - publicationStatus
        - comparabilityGroupKey
      properties:
        id: { type: string }
        clinicId: { type: string }
        outcomeType: { type: string }
        outcomeDefinition: { type: string }
        ratePercentage: { type: number, minimum: 0, maximum: 100 }
        numerator: { type: ["number", "null"], minimum: 0, multipleOf: 1 }
        denominatorCount: { type: ["number", "null"], minimum: 0, multipleOf: 1 }
        denominatorType: { type: string }
        denominatorDefinition: { type: string }
        ageBand: { type: string }
        ageMeasurementPoint: { type: ["string", "null"] }
        eggSource: { type: string }
        treatmentContext: { type: string }
        treatmentType: { type: string }
        priorTreatmentCohort: { type: ["string", "null"] }
        cumulativeMethod: { type: string }
        reportingPeriodStart: { type: string, format: date }
        reportingPeriodEnd: { type: string, format: date }
        yearLabel: { type: string }
        methodologyNotes: { type: string }
        smallSample: { type: boolean }
        source: { $ref: "#/components/schemas/Source" }
        verificationStatus: { type: string }
        publicationStatus: { type: string }
        comparabilityGroupKey: { type: string }
...
    Clinic:
      type: object
      required: [id, slug, name, city, state, services, lastReviewedAt]
      properties:
        id: { type: string }
        slug: { type: string }
        name: { type: string }
        city: { type: string }
        state: { type: string }
        address: { type: ["string", "null"] }
        latitude: { type: ["number", "null"] }
        longitude: { type: ["number", "null"] }
        licensingStatus: { type: ["string", "null"] }
        regulator: { type: ["string", "null"] }
        phone: { type: ["string", "null"] }
        email: { type: ["string", "null"], minLength: 1 }
        website: { type: ["string", "null"], minLength: 1 }
        services:
          type: array
          items: { $ref: "#/components/schemas/Service" }
        headlineObservation:
          oneOf:
            - $ref: "#/components/schemas/RateObservation"
            - type: "null"
        lastReviewedAt: { type: string, format: date }
        demonstrationData: { type: boolean }
    ClinicProfile:
      allOf:
        - $ref: "#/components/schemas/Clinic"
        - type: object
          required: [observations]
          properties:
            observations:
              type: array
              items: { $ref: "#/components/schemas/RateObservation" }
    ClinicListResponse:
      type: object
      required: [items, total, availableLocations, availableServices]
      properties:
        items:
          type: array
          items: { $ref: "#/components/schemas/Clinic" }
        total: { type: number, minimum: 0, multipleOf: 1 }
        availableLocations:
          type: array
          items: { $ref: "#/components/schemas/Location" }
        availableServices:
          type: array
          items: { $ref: "#/components/schemas/Service" }
...
  parameters:
    Search:
      name: q
      in: query
      required: false
      schema: { type: string }
    Location:
      name: location
      in: query
      required: false
      schema: { type: string }
    Service:
      name: service
      in: query
      required: false
      schema: { type: string }
    AgeBand:
      name: ageBand
      in: query
      required: false
      schema: { type: string }
    OutcomeType:
      name: outcomeType
      in: query
      required: false
      schema: { type: string }
    DenominatorType:
      name: denominatorType
      in: query
      required: false
      schema: { type: string }
    EggSource:
      name: eggSource
      in: query
      required: false
      schema: { type: string }
    VerificationType:
      name: verificationType
      in: query
      required: false
      schema: { type: string }
    ReportingYear:
      name: reportingYear
      in: query
      required: false
      schema: { type: number, multipleOf: 1 }
    ClinicSlug:
      name: slug
      in: path
      required: true
      schema: { type: string }
...
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: session