# lib/api-zod/src/generated/types/clinicListResponse.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Evidence-led IVF clinic directory API for fictional demonstration data.
 * OpenAPI spec version: 0.1.0
 */
import type { Clinic } from './clinic';
import type { Location } from './location';
import type { Service } from './service';

export interface ClinicListResponse {
  items: Clinic[];
  /** @minimum 0 */
  total: number;
  availableLocations: Location[];
  availableServices: Service[];
}